backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchaudio
transformers
pyannote.audio
pydub
tqdm
python-dotenv
numpy
sse-starlette==0.10.2
noisereduce==3.0.3
